dev<delete-fixtures-images.sh>: check if directory exists

+++ delete-fixtures-images.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
-rm -rf web/images/items/
-mkdir web/images/items/
+if [ -d /web/images/items ]; then
+    rm -rf web/images/items/
+    mkdir web/images/items/
+fi
